* Begin to change array IDs from gssize to gint
  Use new eh_id_array and sed_cube_water_depth_grid that use gint rather
  than gssize for array IDs.

+++ ew/inflow/inflow_sedflux_interface.c
@@ -259,10 +259,10 @@ inflow_set_bathy_data_from_cube( Sed_cube p , double* width , gssize ind_start ,
 
    if ( p )
    {
-      gssize      len;
-      gssize*     id        = eh_id_array( ind_start , sed_cube_size(p)-1 , &len );
+      gint        len;
+      gint*       id        = eh_id_array (ind_start, sed_cube_size(p)-1, &len);
       double**    bathy     = eh_new_2( double , 3 , sed_cube_size(p) );
-      Eh_dbl_grid g         = sed_cube_water_depth_grid( p , id );
+      Eh_dbl_grid g         = sed_cube_water_depth_grid (p, id);
       double      basin_len;
 
       eh_require( id       );
